Split declaration and definition of FieldValidator and make it as a separate target.
