+++ test/module/shared_model/backend_proto/CMakeLists.txt
@@ -44,6 +44,7 @@ addtest(shared_proto_queries_test
 target_link_libraries(shared_proto_queries_test
     shared_model_proto_backend
     shared_model_ed25519_sha3
+    shared_model_field_validator
     )
 
 addtest(shared_proto_query_responses_test
